requirements.txt
openai~=0.27.6
aiogram~=2.25.1
speech-recognition-fork
python-decouple
gTTS
peewee
openai-whisper
aiolimiter
pytrends
GoogleNews
newspaper3k